# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: CueSeek Agent API
  version: 0.1.0
  description: |
    The API exposed by `cueseekd`, the CueSeek host agent.

    ## Transport and exposure

    This API is **not designed to be exposed to the public internet**. It assumes the
    client can already reach the host over a private network (Tailscale, WireGuard or
    LAN), and performs no TLS termination of its own (ADR-0001).

    Network reachability is not authorisation. Every request except `POST /v1/pair`
    carries a per-device bearer token with explicit scopes (ADR-0006).

    ## Reads are a stream; writes are RPCs

    Live state arrives over `GET /v1/stream`: a full snapshot as the first event, then
    deltas. Reconnecting yields a new full snapshot — there is no replay buffer, because
    the state is small and correctness beats bandwidth.

    `GET /v1/services` exists for cold start and for clients that cannot hold a stream
    open. Wear OS uses it exclusively: a held connection does not survive Doze, so
    nothing background-critical may depend on the stream.

    ## Actions are asynchronous

    Actions return `202 Accepted` with an action id; progress and terminal state arrive
    over the stream. This is not ceremony — it is forced by the platform. `RestartUnit`
    over D-Bus returns a *job* immediately and completes later, so the agent has no
    synchronous result to report even if it wanted one. Host power actions are worse
...

    ## Capabilities

...
  contact:
...
  license:
...
servers:
...
security:
  - deviceToken: []
paths:
...
  /v1/services/{serviceId}/actions/{actionId}:
    post:
      operationId: invokeServiceAction
      tags: [services]
      summary: Invoke an action on a service
      description: |
        Accepts the action and returns immediately with an action id. Progress and
        terminal state arrive over the stream.

        The agent cannot report a synchronous result: `RestartUnit` returns a systemd
        *job* that completes later, and M0 confirmed that a `rc=0` return means only
        that the job was enqueued (`docs/m0-findings.md`).

        Requires the `service.control` scope. Scope enforcement happens here, in the
        agent — client-side confirmation prompts are user experience, not a control.
      x-required-scope: service.control
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/ActionId"
      responses:
        "202":
          description: Action accepted. Watch the stream for its outcome.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The action is not currently possible, e.g. already in progress.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: |
        A per-device token issued at pairing time. Tokens are stored by the agent as
        hashes only, are individually revocable, and carry independent scopes:

        | Scope | Grants |
        | --- | --- |
        | `read` | Read state: system, devices, services, the stream. |
        | `service.control` | Invoke service actions, e.g. restart Jellyfin. |
        | `devices.manage` | Revoke paired devices. |
        | `host.power` | Reboot or shut down the machine. |

        Scopes are separate grants, not tiers, and the set is chosen so that the
        dangerous powers can be withheld individually. A paired watch is typically issued
        `read` and `service.control` only: it can restart Jellyfin, but is structurally
        incapable of powering off the server or of locking other devices out — regardless
        of what its UI offers.

        The scope each operation requires is given by `x-required-scope`. It is an
        extension rather than a formal OpenAPI construct because scopes are only modelled
        natively for OAuth2 and OpenID Connect, and neither applies here (ADR-0006).
  parameters:
...
    ServiceId:
      name: serviceId
      in: path
      required: true
      description: Stable service identifier, e.g. `jellyfin`.
      schema: { type: string }
    ActionId:
      name: actionId
      in: path
      required: true
      description: Action identifier as advertised in the service's `actions` list.
      schema: { type: string }
  responses:
...
    Unauthorized:
      description: Missing, invalid or revoked token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The token is valid but lacks the scope this operation requires.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: |
        Error payload following RFC 9457 (problem+json). One error shape across the whole
        API means clients write one error mapper rather than one per endpoint.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type.
          example: "https://cueseek.dev/problems/insufficient-scope"
        title:
          type: string
          description: Short, human-readable summary. Stable for a given type.
          example: "Insufficient scope"
        status:
          type: integer
          description: HTTP status code.
          example: 403
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this specific occurrence.

    # ---------------------------------------------------------------- system
...
    ActionStatus:
      type: string
      enum: [pending, running, succeeded, failed]
    ActionAccepted:
      type: object
      required: [action_id, service_id, action, status, accepted_at]
      properties:
        action_id:
          type: string
          description: Correlates this invocation with `action_progress` stream events.
        service_id: { type: string }
        action: { type: string }
        status: { $ref: "#/components/schemas/ActionStatus" }
        accepted_at: { type: string, format: date-time }
...